comments fixes not sure if I really did DI correctly in the controllers, but I think I did
